requirements.txt
accelerate>=0.20.0
appdirs
loralib
bitsandbytes>=0.41.0
black
black[jupyter]
datasets>=2.0.0
fire
peft>=0.4.0
transformers>=4.31.0
sentencepiece
gradio>=3.0.0
wandb>=0.15.0
torch>=2.0.0
einops
tiktoken
safetensors>=0.3.0
azureml-mlflow
# xFormers for memory efficient attention (可选，unsloth 优先使用 flash-attn)
# xformers
# Optimized kernels
triton

# ========== Unsloth + Flash Attention ==========
# TRL for SFTTrainer (unsloth 需要)
trl>=0.7.0